bump: Upgrade to fmt 8.1.1

+++ src/core/CacheEntryHeader.cpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2019-2021 Joel Rosdahl and other contributors
+// Copyright (C) 2019-2022 Joel Rosdahl and other contributors
 //
 // See doc/AUTHORS.adoc for a complete list of contributors.
 //
@@ -74,7 +74,10 @@ CacheEntryHeader::dump(FILE* const stream) const
 {
   PRINT(stream, "Magic: {:04x}\n", magic);
   PRINT(stream, "Entry format version: {}\n", entry_format_version);
-  PRINT(stream, "Entry type: {} ({})\n", entry_type, to_string(entry_type));
+  PRINT(stream,
+        "Entry type: {} ({})\n",
+        static_cast<uint8_t>(entry_type),
+        to_string(entry_type));
   PRINT(stream,
         "Compression type: {}\n",
         compression::type_to_string(compression_type));
